saveMemory 메소드 동기화

+++ src/test/java/me/programmers/calculator/engine/model/HistoryTest.java
@@ -2,7 +2,7 @@ package me.programmers.calculator.engine.model;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class HistoryTest {
 
@@ -11,10 +11,10 @@ class HistoryTest {
     @Test
     void saveMemoryTest() {
         String problem = "1+2";
-        int sum = 3;
+        long sum = 3;
 
         history.saveMemory(problem, sum);
 
-        assertEquals(history.getMemory().get(1), problem+"="+sum);
+        assertEquals(history.getMemory().get(1), problem + "=" + sum);
     }
 }
